api/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
lyricsgenius>=3.0.0
nltk>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
